Create components directory, Create app directory inside components, Move app js and css files into app directory
